Interactables getestet (Unit test), Ausgaben erweitert. --> Interactables können nun bei Rundenende mit Funktion: CreateRoundEndInteractables() gespawnt werden, oder auch einzelne Währungen mit SpawnWährung() // Heal mit SpawnHeal() --> Entity/Collidable Delete funktion muss noch eingefügt werden, dass Gegner auch z. B. Währung fallen lässt!

+++ ComputergrafikSpiel/Model/Entity/CreateInteractable.cs
@@ -14,7 +14,7 @@ namespace ComputergrafikSpiel.Model.Entity
     {
         private readonly string texturename;
 
-        public CreateInteractable(PlayerEnum.Stats stats, int positionX, int positionY)
+        public CreateInteractable(PlayerEnum.Stats stats, float positionX, float positionY)
         {
             switch (stats)
             {
